requirements.txt
celery==5.0.0
Django==3.1.1
djangorestframework==3.11.0
redis==3.5.3
requests==2.24.0
